+++ Flight Algorithms/data_collection.py
@@ -5,13 +5,22 @@ Created on Sat Mar  4 12:22:46 2023
 
 @author: zrummler
 """
+import sys
 import pandas as pd
 import numpy as np
 
+sys.path.append('../Data Generation')
+
 # code for simulation on PC
 
+imu_reading_number = 0
+gps_reading_number = 0
+last_time = 0
+
+GPS_UPDATE_TIME = 1
+
 # reading IMU data
-file_data = pd.read_csv("traj_raster_30mins_20221115_160156.csv")
+file_data = pd.read_csv("../Data Generation/traj_raster_30mins_20221115_160156.csv")
 imu_t_sec = np.array(file_data["t [sec]"])
 accel_X = np.array(file_data["accel_X [m/s^2]"])
 accel_Y = np.array(file_data["accel_Y [m/s^2]"])
@@ -21,7 +30,7 @@ gyro_Y = np.array(file_data["gyro_X [rad/s]"])
 gyro_Z = np.array(file_data["gyro_X [rad/s]"])
     
 # reading GPS data
-file_data = pd.read_csv("gps_and_barometer.csv")
+file_data = pd.read_csv("../Data Generation/gps_and_barometer.csv")
 gps_t_sec = np.array(file_data["t [sec]"])
 gps_lat = np.array(file_data["GPS_lat"])
 gps_long = np.array(file_data["GPS_long"])
@@ -29,7 +38,6 @@ gps_alt = np.array(file_data["GPS_alt"])
 
 
 """ returns the next IMU reading (simulation) """
-imu_reading_number = 0
 def get_next_imu_reading():
     global imu_reading_number
 
@@ -49,7 +57,6 @@ def get_next_imu_reading():
     
 
 """ returns the next GPS reading (simulation) """
-gps_reading_number = 0
 def ping_gps_for_reading():
     global gps_reading_number
     
@@ -69,11 +76,10 @@ def ping_gps_for_reading():
     return lat, long, alt, dt
     
 """ how to check if the GPS is ready for next data """
-last_time = 0
 def gps_is_ready(running_time):
     global last_time
     
-    if (running_time - last_time) >= 1:
+    if (running_time - last_time) >= GPS_UPDATE_TIME:
         last_time = running_time
         return True
